# Micro-op for interleave_evenly