### Issue reference

Closes #701

### Changes

- Micro-optimization for **interleave_evenly**

### Implementation Details

Use `map` instead of `zip` as `map` is slightly faster for applying `operator.sub`.

Benchmarked as follows

**OS:** Linux x64 **CPU:** AMD Ryzen 7 5800X

```python
from more_itertools import interleave_evenly
%timeit list(interleave_evenly(([x for x in range(26)] * 100, [chr(ord('a') + x) for x in range(26)] * 100)))

Before: 1.1 ms ± 12 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
After : 895 µs ± 13.5 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```