@bbayles 2023-03-23
Usually this sort of thing involves a performance hit on PyPy. Did you test that too?