@elliotwutingfeng 2023-03-23
Did a couple of benchmarks. It's an improvement on PyPy. On a Raspberry Pi it was a little slower.

**OS:** Linux x64 **CPU:** AMD Ryzen 7 5800X **Python Version:** PyPy 3.9.16

```python
Before: 362 µs ± 3.37 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
After : 317 µs ± 1.82 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

**OS:** Raspberry Pi OS 5.15.61 **CPU:** Raspberry Pi 3 Model B **Python Version:** Python 3.9.2
 
```python
Before: 49.7 ms ± 2.82 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
After : 51 ms ± 134 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```